pyproject.toml
[project]
name = "agent-weather-info"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.2.9",
    "python-decouple>=3.8",
]
